fix use-before-init bug
